apps/api/requirements.txt
Django==5.1
djangorestframework==3.15
djangorestframework-gis==1.0
psycopg2-binary==2.9
django-cors-headers==4.4
django-environ==0.11
celery==5.4
redis==5.1
meilisearch==0.31
requests==2.31
openai==1.12
sentence-transformers==2.3
pgvector==0.2
